requirements.txt
Flask==1.1.2
Flask-SocketIO==4.3.1
google-api-python-client==1.12.5
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
gunicorn==20.0.4
Jinja2==2.11.2
gevent==20.9.0
gevent-websocket==0.10.1